+++ src/main/java/com/zuniorteam/racingcar/RacingGameBoard.java
@@ -2,8 +2,8 @@ package com.zuniorteam.racingcar;
 
 import com.sun.istack.internal.NotNull;
 import com.zuniorteam.racingcar.core.RacingGame;
-import com.zuniorteam.racingcar.core.dto.GameInput;
-import com.zuniorteam.racingcar.core.dto.GameResult;
+import com.zuniorteam.racingcar.dto.GameInput;
+import com.zuniorteam.racingcar.dto.GameResult;
 import com.zuniorteam.racingcar.core.strategy.MovingStrategy;
 import com.zuniorteam.racingcar.view.GameInputView;
 import com.zuniorteam.racingcar.view.GameResultView;
