+++ src/main/java/com/zuniorteam/racingcar/view/GameInputView.java
@@ -1,6 +1,6 @@
 package com.zuniorteam.racingcar.view;
 
-import com.zuniorteam.racingcar.core.dto.GameInput;
+import com.zuniorteam.racingcar.dto.GameInput;
 
 public interface GameInputView {
     GameInput listen();
