+++ src/main/java/com/zuniorteam/racingcar/view/GameResultView.java
@@ -1,6 +1,6 @@
 package com.zuniorteam.racingcar.view;
 
-import com.zuniorteam.racingcar.core.dto.GameResult;
+import com.zuniorteam.racingcar.dto.GameResult;
 
 public interface GameResultView {
     void draw(GameResult gameResult);
