+++ src/test/java/com/zuniorteam/racingcar/dto/GameResultTest.java
@@ -1,4 +1,4 @@
-package com.zuniorteam.racingcar.core.dto;
+package com.zuniorteam.racingcar.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,19 +8,19 @@ import java.util.Collections;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class StepResultTest {
 
+class GameResultTest {
 
     @DisplayName("New")
     @Test
     void testNewInstance(){
-        assertDoesNotThrow(() -> new StepResult(Collections.emptyList()));
+        assertDoesNotThrow(() -> new GameResult(Collections.emptyList()));
     }
 
     @DisplayName("New, Assertion Error")
     @Test
     void testNewInstance02(){
-        assertThrows(AssertionError.class, () -> new StepResult(null));
+        assertThrows(AssertionError.class, () -> new GameResult(null));
     }
 
 }
